Add the ability to login as another user and add multiple domains into itself. For example to add multiple domain for a user muser use the following format in ENV file: ===== DA_USER_USERNAME2="muser" DA_USER_USERPASS2="password" DA_USER_EMAIL2="webmaster@kalibmedia.com" DA_USER_DOMAIN2="kalibmedia.com kalibtv.com kalibtest.com" DA_USER_PACKAGE_NAME2="USER_UNLIMITED" DA_USER_DOMAIN_ENABLE_SSL2="yes" =====
The script will automatically login as user 'muser' and add the 3 domains for itself

Now the API authentication can be passed dynamically. For example for this user use the following syntax:

start+${role_user}+${role_password}

I use the symbol + instead of _ because '+' is a unique and useless character which cannot be used as password in DA.

The previous API auth call has changed. eg: from start_admin change to start+admin